third_party/blink/web_tests/platform/generic/external/wpt/editing/other/restoration-expected.txt

This is a testharness.js-based test.
PASS insertparagraph  starting bold
PASS insertparagraph  starting not bold
PASS insertlinebreak  starting bold
PASS insertlinebreak  starting not bold
FAIL delete  starting bold assert_equals: bold state expected false but got true
FAIL delete  starting not bold assert_equals: bold state expected true but got false
FAIL forwarddelete  starting bold assert_equals: bold state expected false but got true
FAIL forwarddelete  starting not bold assert_equals: bold state expected true but got false
FAIL insertorderedlist  starting bold assert_equals: bold state expected false but got true
FAIL insertorderedlist  starting not bold assert_equals: bold state expected true but got false
FAIL insertunorderedlist  starting bold assert_equals: bold state expected false but got true
FAIL insertunorderedlist  starting not bold assert_equals: bold state expected true but got false
FAIL indent  starting bold assert_equals: bold state expected false but got true
FAIL indent  starting not bold assert_equals: bold state expected true but got false
PASS outdent  starting bold
PASS outdent  starting not bold
FAIL justifyleft  starting bold assert_equals: bold state expected false but got true
FAIL justifyleft  starting not bold assert_equals: bold state expected true but got false
FAIL justifyright  starting bold assert_equals: bold state expected false but got true
FAIL justifyright  starting not bold assert_equals: bold state expected true but got false
FAIL justifycenter  starting bold assert_equals: bold state expected false but got true
FAIL justifycenter  starting not bold assert_equals: bold state expected true but got false
FAIL justifyfull  starting bold assert_equals: bold state expected false but got true
FAIL justifyfull  starting not bold assert_equals: bold state expected true but got false
FAIL formatblock div starting bold assert_equals: bold state expected false but got true
FAIL formatblock div starting not bold assert_equals: bold state expected true but got false
FAIL formatblock blockquote starting bold assert_equals: bold state expected false but got true
FAIL formatblock blockquote starting not bold assert_equals: bold state expected true but got false
PASS inserthorizontalrule  starting bold
FAIL inserthorizontalrule  starting not bold assert_equals: bold state expected true but got false
PASS insertimage a starting bold
FAIL insertimage a starting not bold assert_equals: bold state expected true but got false
PASS inserttext bar starting bold
PASS inserttext bar starting not bold
Harness: the test ran to completion.
